// src/common/pagination/interfaces/paginated.interface.ts
export interface PaginatedInterface<T> {
  data: T[];
  meta: {
    totalItems: number;
    itemCount: number;
    itemsPerPage: number;
    totalPages: number;
    currentPage: number;
  };
  links: {
    first: string;
    previous: string;
    next: string;
    last: string;
    current: string;
  };
}